+++ miniapp/CMakeLists.txt
@@ -1,5 +1,3 @@
-set(HEADERS
-)
 set(MINIAPP_SOURCES
     miniapp.cpp
     io.cpp
@@ -7,15 +5,8 @@ set(MINIAPP_SOURCES
     morphology_pool.cpp
     trace.cpp
 )
-set(MINIAPP_SOURCES_CUDA
-    miniapp.cu
-    io.cpp
-    miniapp_recipes.cpp
-    morphology_pool.cpp
-    trace.cpp
-)
 
-add_executable(miniapp.exe ${MINIAPP_SOURCES} ${HEADERS})
+add_executable(miniapp.exe ${MINIAPP_SOURCES})
 
 target_link_libraries(miniapp.exe LINK_PUBLIC ${NESTMC_LIBRARIES})
 target_link_libraries(miniapp.exe LINK_PUBLIC ${EXTERNAL_LIBRARIES})
@@ -25,7 +16,8 @@ if(NMC_WITH_MPI)
     set_property(TARGET miniapp.exe APPEND_STRING PROPERTY LINK_FLAGS "${MPI_C_LINK_FLAGS}")
 endif()
 
-set_target_properties(miniapp.exe
-   PROPERTIES
-   RUNTIME_OUTPUT_DIRECTORY "${CMAKE_BINARY_DIR}/miniapp"
+set_target_properties(
+    miniapp.exe
+    PROPERTIES
+    RUNTIME_OUTPUT_DIRECTORY "${CMAKE_BINARY_DIR}/miniapp"
 )
