--- miniapp/miniapp.cu
@@ -1,1 +1,0 @@
-#include "miniapp.cpp"
