+++ src/backends/gpu/matrix_state_flat.hpp
@@ -1,17 +1,34 @@
 #pragma once
 
+#include <backends/fvm_types.hpp>
 #include <memory/memory.hpp>
 #include <memory/wrappers.hpp>
 #include <util/span.hpp>
 #include <util/partition.hpp>
 #include <util/rangeutil.hpp>
 
-#include "kernels/solve_matrix.hpp"
-#include "kernels/assemble_matrix.hpp"
-
 namespace nest {
 namespace mc {
 namespace gpu {
+
+void solve_matrix_flat(
+    fvm_value_type* rhs,
+    fvm_value_type* d,
+    const fvm_value_type* u,
+    const fvm_size_type* p,
+    const fvm_size_type* cell_cv_divs,
+    int num_mtx);
+
+void assemble_matrix_flat(
+    fvm_value_type* d,
+    fvm_value_type* rhs,
+    const fvm_value_type* invariant_d,
+    const fvm_value_type* voltage,
+    const fvm_value_type* current,
+    const fvm_value_type* cv_capacitance,
+    const fvm_size_type* cv_to_cell,
+    const fvm_value_type* dt_cell,
+    unsigned n);
 
 /// matrix state
 template <typename T, typename I>
@@ -95,25 +112,16 @@ struct matrix_state_flat {
     //   voltage [mV]
     //   current [nA]
     void assemble(const_view dt_cell, const_view voltage, const_view current) {
-        // determine the grid dimensions for the kernel
-        auto const n = voltage.size();
-        auto const block_dim = 128;
-        auto const grid_dim = impl::block_count(n, block_dim);
-
-        assemble_matrix_flat<value_type, size_type><<<grid_dim, block_dim>>> (
+        // perform assembly on the gpu
+        assemble_matrix_flat(
             d.data(), rhs.data(), invariant_d.data(), voltage.data(),
             current.data(), cv_capacitance.data(), cv_to_cell.data(), dt_cell.data(), size());
     }
 
     void solve() {
-        // determine the grid dimensions for the kernel
-        auto const block_dim = 128;
-        auto const grid_dim = impl::block_count(num_matrices(), block_dim);
-
         // perform solve on gpu
-        solve_matrix_flat<value_type, size_type><<<grid_dim, block_dim>>> (
-            rhs.data(), d.data(), u.data(), parent_index.data(),
-            cell_cv_divs.data(), num_matrices());
+        solve_matrix_flat(rhs.data(), d.data(), u.data(), parent_index.data(),
+                          cell_cv_divs.data(), num_matrices());
     }
 
     std::size_t size() const {
