+++ tests/unit/test_matrix.cu
@@ -12,6 +12,10 @@
 #include <memory/memory.hpp>
 #include <util/span.hpp>
 
+#include <backends/gpu/matrix_state_flat.hpp>
+#include <backends/gpu/matrix_state_interleaved.hpp>
+#include <backends/gpu/kernels/interleave.hpp>
+
 #include <cuda.h>
 
 using namespace nest::mc;
