--- tests/validation/validate_soma.cu
@@ -1,9 +1,0 @@
-#include "validate_soma.hpp"
-
-#include "../gtest.h"
-
-const auto backend = nest::mc::backend_kind::gpu;
-
-TEST(soma, numeric_ref) {
-    validate_soma(backend);
-}
